requirements.txt
opencv-python
numpy
tensorflow
Pillow
asyncio
psutil
python-dotenv